requirements.txt
# TimeToShopping_bot Dependencies

# Core bot framework
aiogram==3.13.0
aiohttp==3.9.1

# AI integration
openai==1.52.0

# Database
aiosqlite==0.20.0
sqlalchemy==2.0.25
alembic==1.13.1

# Task scheduling
apscheduler==3.10.4

# Configuration and environment
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Date/time utilities
python-dateutil==2.8.2

# CSV export
pandas==2.1.4

# Async utilities
asyncio-throttle==1.0.2

# Type hints
typing-extensions==4.9.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0